+++ lib/CssSyntaxGenerator.js
@@ -1,7 +1,15 @@
 const Generator = require('chance-generators/lib/Generator');
+const RangeGenerator = require('./RangeGenerator');
 const cssSyntaxParser = require('css-syntax-parser').default;
-const defaultGenerators = require('./defaultGenerators');
-const { InvalidSyntaxError } = require('./utils');
+const _ = require('lodash');
+
+class InvalidSyntaxError extends Error {
+  constructor(node) {
+    super(
+      `Don't know how to generate ${JSON.stringify(node, undefined, '  ')}`
+    );
+  }
+}
 
 class CssSyntaxGenerator extends Generator {
   constructor(nodeOrSyntax, { descriptors = {} } = {}) {
@@ -29,36 +37,141 @@ class CssSyntaxGenerator extends Generator {
     }
   }
 
-  static setGenerators(generators = {}) {
-    CssSyntaxGenerator.generators = { ...defaultGenerators, ...generators };
-  }
-
   generateLiteral(chance) {
-    return CssSyntaxGenerator.generators.generateLiteral.bind(this)(chance);
+    const { node } = this;
+    return new RangeGenerator(node._value, node.range).generate(chance);
   }
 
   generateKeyword(chance) {
-    return CssSyntaxGenerator.generators.generateKeyword.bind(this)(chance);
+    const { node } = this;
+    return new RangeGenerator(node._value, node.range).generate(chance);
   }
 
   generateString(chance) {
-    return CssSyntaxGenerator.generators.generateString.bind(this)(chance);
+    const { node } = this;
+    return new RangeGenerator(node._value, node.range).generate(chance);
   }
 
   generateBrackets(chance) {
-    return CssSyntaxGenerator.generators.generateBrackets.bind(this)(chance);
+    const {
+      node,
+      options: { descriptors },
+    } = this;
+
+    if (node.multiplier === '#') {
+      return _.range(chance.integer({ min: 1, max: 5 }))
+        .map((i) =>
+          new CssSyntaxGenerator(node.content, { descriptors }).generate(chance)
+        )
+        .join(', ');
+    } else if (node.multiplier === '!') {
+      let value;
+      do {
+        value = new CssSyntaxGenerator(node.content, {
+          descriptors,
+        }).generate(chance);
+      } while (value.trim() === '');
+      return value;
+    } else {
+      return new CssSyntaxGenerator(node.content, { descriptors }).generate(
+        chance
+      );
+    }
   }
 
   generateDataType(chance) {
-    return CssSyntaxGenerator.generators.generateDataType.bind(this)(chance);
+    const CssNamedSyntaxGenerator = require('./CssNamedSyntaxGenerator');
+    const {
+      node,
+      options: { descriptors },
+    } = this;
+
+    let dataTypeGenerator;
+    if (node.nonTerminal) {
+      const CssDeclarationValueGenerator = require('./CssDeclarationValueGenerator');
+      dataTypeGenerator = new CssDeclarationValueGenerator({
+        propertyName: node.name,
+      });
+    } else if (descriptors[node.name]) {
+      dataTypeGenerator = new CssSyntaxGenerator(
+        descriptors[node.name].syntax,
+        {}
+      );
+    } else {
+      dataTypeGenerator = new CssNamedSyntaxGenerator(node.name);
+    }
+    return new RangeGenerator(dataTypeGenerator, node.range).generate(chance);
   }
 
   generateComposed(chance) {
-    return CssSyntaxGenerator.generators.generateComposed.bind(this)(chance);
+    const {
+      node,
+      options: { descriptors },
+    } = this;
+
+    if (node.combinator === '||') {
+      return chance
+        .pickset(
+          node.children,
+          chance.integer({ min: 1, max: node.children.length })
+        )
+        .map((child) =>
+          new CssSyntaxGenerator(child, { descriptors }).generate(chance)
+        )
+        .join(' ');
+    } else if (node.combinator === '&&') {
+      return chance
+        .shuffle(node.children)
+        .map((child) =>
+          new CssSyntaxGenerator(child, { descriptors }).generate(chance)
+        )
+        .join(' ');
+    } else if (node.combinator === '|') {
+      return new CssSyntaxGenerator(chance.pickone(node.children), {
+        descriptors,
+      }).generate(chance);
+    } else if (node.combinator === ' ') {
+      const tokens = [];
+      // Generate the children, but omit commas that don't satisfy
+      // the criteria in https://drafts.csswg.org/css-values-3/#component-types
+      let state = 'INITIAL';
+      for (const child of node.children) {
+        if (child.type === 'literal' && child._value === ',') {
+          if (state === 'AFTER_CONTENT') {
+            state = 'MAYBE_COMMA';
+          }
+        } else {
+          const token = new CssSyntaxGenerator(child, {
+            descriptors,
+          }).generate(chance);
+          if (token.trim() !== '') {
+            if (state === 'MAYBE_COMMA') {
+              tokens.push(',');
+            }
+            tokens.push(token);
+            state = 'AFTER_CONTENT';
+          }
+        }
+      }
+
+      return tokens.join(' ');
+    }
+
+    throw new InvalidSyntaxError();
   }
 
   generateMethod(chance) {
-    return CssSyntaxGenerator.generators.generateMethod.bind(this)(chance);
+    const {
+      node,
+      options: { descriptors },
+    } = this;
+    if (node.params) {
+      return `${node.name}(${new CssSyntaxGenerator(node.params, {
+        descriptors,
+      }).generate(chance)})`;
+    } else {
+      return `${node.name}()`;
+    }
   }
 
   // Wrap all _generate calls to compensate for https://github.com/zanomate/css-syntax-parser/issues/1
@@ -83,7 +196,6 @@ class CssSyntaxGenerator extends Generator {
 
   _generate(chance) {
     const { node } = this;
-
     if (node.multiplier === '?' && chance.bool()) {
       return '';
     }
@@ -103,11 +215,9 @@ class CssSyntaxGenerator extends Generator {
       case 'method':
         return this.generateMethod(chance);
       default:
-        throw new InvalidSyntaxError(node);
+        throw new InvalidSyntaxError();
     }
   }
 }
-
-CssSyntaxGenerator.setGenerators();
 
 module.exports = CssSyntaxGenerator;
