+++ lib/index.js
@@ -8,7 +8,6 @@ const CssAtRuleGenerator = require('./CssAtRuleGenerator');
 const CssStyleRuleGenerator = require('./CssStyleRuleGenerator');
 const CssStylesheetGenerator = require('./CssStylesheetGenerator');
 const createGeneratorFacade = require('chance-generators/lib/createGeneratorFacade');
-const defaultGenerators = require('./defaultGenerators');
 
 module.exports = {
   syntax: createGeneratorFacade(CssSyntaxGenerator),
@@ -20,5 +19,4 @@ module.exports = {
   atRule: createGeneratorFacade(CssAtRuleGenerator),
   styleRule: createGeneratorFacade(CssStyleRuleGenerator),
   stylesheet: createGeneratorFacade(CssStylesheetGenerator),
-  defaultGenerators,
 };
