+++ test/CssSyntaxGenerator.js
@@ -5,7 +5,6 @@ const CssSyntaxGenerator = require('../lib/CssSyntaxGenerator');
 
 describe('CssSyntaxGenerator', () => {
   beforeEach(() => chanceCache.clear());
-  afterEach(() => CssSyntaxGenerator.setGenerators());
 
   it('supports an already parsed node', () => {
     expect(
@@ -115,17 +114,5 @@ describe('CssSyntaxGenerator', () => {
       'foo(-4813963569135616)',
       'foo(-5410196146880512)',
     ]);
-  });
-
-  it('should support custom generator methods', function () {
-    CssSyntaxGenerator.setGenerators({
-      generateKeyword() {
-        return 'Lorem ipsum';
-      },
-    });
-
-    expect(new CssSyntaxGenerator('string').first()).toEqualSnapshot(
-      'Lorem ipsum'
-    );
   });
 });
